+++ .vscode/settings.json
@@ -8,6 +8,7 @@
         "applocale",
         "daos",
         "dartz",
+        "datasources",
         "dont",
         "endregion",
         "hmac",
