+++ src/sprightly/lib/data/constants/enums.dart
