+++ src/sprightly/lib/data/repositories/activities.dart
@@ -1,9 +1,9 @@
 library sprightly.repositories;
 
 import 'package:sprightly/core/exceptions.dart';
-import 'package:sprightly/models/constants/enums.dart';
-import 'package:sprightly/models/dao.dart';
-import 'package:sprightly/models/db/database.dart';
+import 'package:sprightly/data/constants/enums.dart';
+import 'package:sprightly/data/dao.dart';
+import 'package:sprightly/data/datasources/database.dart';
 
 mixin _BaseData {
   SystemDao _dao;
