+++ src/sprightly/lib/data/repositories/settings.dart
@@ -2,10 +2,10 @@ library sprightly.repositories;
 
 import 'package:sprightly/core/exceptions.dart';
 import 'package:sprightly/extensions/enum_extensions.dart';
-import 'package:sprightly/models/constants/enums.dart';
-import 'package:sprightly/models/constants/values.dart';
-import 'package:sprightly/models/dao.dart';
-import 'package:sprightly/models/db/database.dart';
+import 'package:sprightly/data/constants/enums.dart';
+import 'package:sprightly/data/constants/values.dart';
+import 'package:sprightly/data/dao.dart';
+import 'package:sprightly/data/datasources/database.dart';
 import 'package:sprightly/core/app_parameter.dart';
 
 mixin _BaseData {
