+++ src/sprightly/lib/features/settings/data/datasources/README.md
@@ -1,5 +1,5 @@
 Already handled by 
 
 ```dart
-import 'package:sprightly/models/db/database.dart';
+import 'package:sprightly/data/datasources/database.dart';
 ```
