+++ src/sprightly/lib/features/themes/data/models/README.md
@@ -1,5 +1,5 @@
 Already handled by 
 
 ```dart
-import 'package:sprightly/models/db/database.g.dart';
+import 'package:sprightly/data/datasources/database.g.dart';
 ```
